The First commitafter run program

+++ Program.cs
@@ -14,7 +14,7 @@ Console.Write("Please enter your age: ");
 age = Convert.ToInt32(Console.ReadLine());
 
 Console.Write("Please enter your salary: ");
-salary = Convert.ToInt32(Console.ReadLine());
+salary = Convert.ToDecimal(Console.ReadLine());
 
 Console.Write("Please enter your gender (M or F): ");
 gender = Convert.ToChar(Console.ReadLine());
